Move CacheContent and DynamicContent to Cache folder
